+++ class/GaiaAlpha/Cli/VendorCommands.php
@@ -26,6 +26,7 @@ class VendorCommands
             'https://unpkg.com/leaflet@1.9.4/dist/images/marker-icon-2x.png' => $imagesDir . '/marker-icon-2x.png',
             'https://unpkg.com/leaflet@1.9.4/dist/images/marker-shadow.png' => $imagesDir . '/marker-shadow.png',
             'https://unpkg.com/vue@3.3.4/dist/vue.esm-browser.js' => $rootDir . '/www/js/vendor/vue.esm-browser.js',
+            'https://unpkg.com/globe.gl@2.28.0/dist/globe.gl.min.js' => $rootDir . '/www/js/vendor/globe.gl.js',
         ];
 
         foreach ($files as $url => $path) {